requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0